Add kallsyms_info support
Virtually all desktop & server distribution kernels are built with
kallsyms support, and with kallsyms containing all kernel
symbols (CONFIG_KALLSYMS_ALL), [1]. Since the kernel commits
5fd8fea935a10 ("vmcoreinfo: include kallsyms symbols") and
f09bddbd86619 ("vmcoreinfo: add kallsyms_num_syms symbol") in kernel
6.0, the vmcoreinfo note has included all symbols which are necessary
for an external tool to parse and understand kallsyms from a vmcore,
without additional debuginfo.

As a practical consideration, this unlocks the ability to use any kernel
symbol in makedumpfile, without:

(a) needing to coordinate its inclusion in the vmcoreinfo, or backport
such kernel commits,
(b) needing to use the DWARF debuginfo, or
(c) worrying about the available space of the vmcoreinfo (which is
implemented as a PAGE_SIZE memory region, but could be expanded if
necessary).

In other words, it is suitable for use by most, if not all, desktop and
server distributions without any extra effort. Introduce support for
parsing kallsyms in kernels with the necessary symbols available. This
support code is modified and simplified from the drgn debugger.

This implementation contains support for the primary method of encoding
addresses, the base-relative one, which is the only one currently
present in upstream kernels. However, in older kernel versions there
were two other encodings: absolute, and "absolute per-cpu". The former
was less commonly used, but the latter was the primary dominant encoding
method for x86_64 prior to commit 01157ddc58dc2 ("kallsyms: Remove
KALLSYMS_ABSOLUTE_PERCPU") in kernel 6.15. We support all three methods.

[1]: https://oracle.github.io/kconfigs/?config=UTS_RELEASE&config=KALLSYMS_ALL

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>