makedumpfile: add --retain-user-stacks
This implements the ability to retain anonymous pages that are
associated with userspace stacks. This allows the resulting vmcores to
contain enough information to produce userspace stack traces, which can
be useful for debugging. There are many caveats that should be
mentioned:

1. A task_struct (userspace "thread") is not guaranteed to have a single
stack. Processes may do stack switching, meaning that there are multiple
stacks in use by a thread at a time, e.g. when handling coroutines or
other userspace threading. There's not even any guarantee that a thread
uses even one stack, although it would be surprising if it did not.
2. Further, the kernel is not informed of which memory region(s) the
application considers its stack(s). While the kernel tends to create an
initial stack VMA with VM_GROWS{UP,DOWN} on exec(), it's common for user
applications (typically, glibc) to create a new mmapped stack region
with no special flags.
3. We have no guarantees about the memory region that a process may wish
to use for its stack. It's common for each stack to reside in its own
VMA separate from heap or other memory regions. But a task could place
its stack into a shared memory region, or even into a file-backed
region.

To simplify things, this implementation optimizes for the common case.
We use the stack pointer stored by the kernel at the top of its stack,
and we assume this is the top of the user stack. Further, we only try to
handle stack VMAs which have exactly anon_vma which is not shared to any
other mm. We assume the stack grows up, and we try to include all memory
up to the end of the VMA, or up to a limit so we don't end up retaining
too many pages in the vmcore.

Thus far, the implementation only supports x86_64, and it is enabled by
the flag "--retain-user-stacks". It seems like this feature should be
somehow enabled via the dump level argument, but doing so in a backward
compatible way is a head-scratcher, so I've punted on that for now.

Signed-off-by: Stephen Brennan <stephen.s.brennan@oracle.com>